+++ includes/utils_c3d.h
@@ -6,7 +6,7 @@
 /*   By: mbos <mbos@student.le-101.fr>              +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/12/26 13:08:50 by melodiebos   #+#   ##    ##    #+#       */
-/*   Updated: 2020/01/16 12:52:15 by mbos        ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/01/16 16:24:47 by mbos        ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -22,10 +22,10 @@ int			color_to_int(t_color c);
 int			get_str_int_len(char *str);
 int			occur_in_str(char c,char *line);
 int			get_tile(t_map *m, int x, int y);
-void		rotate_vect(t_vect *v, float angle);
+void		rotate_vect(t_vect *v, double angle);
 t_bool		utils_there_is_a_wall_at(t_rays *r, t_vect v, t_mlx *mlx);
-float		utils_dist_obstacle_cam(t_vect obs, t_vect cam_pos);
-float		degrees_to_radian(float angle_in_degrees);
+double		utils_dist_obstacle_cam(t_vect obs, t_vect cam_pos);
+double		degrees_to_radian(double angle_in_degrees);
 void		ft_pixel_put(t_mlx *mlx, int x, int y, int color);
 void		draw_pix(t_mlx *mlx, t_vect v, int size_obj);
